+++ public/js/models/network-pools.js
@@ -1,5 +1,7 @@
 var Backbone = require('backbone');
 var NetworkPool = require('./network-pool');
+
+var Collection = require('./collection');
 /*
     GET /network_pools/3b5913ec-42e6-4803-9c0b-c9b1c5603520
     {
@@ -12,7 +14,7 @@ var NetworkPool = require('./network-pool');
     }
 */
 
-module.exports = Backbone.Collection.extend({
+module.exports = Collection.extend({
     model: NetworkPool,
     url: "/_/network_pools"
 });